Implement authentication for some endpoints
Move swagger documentation to interfaces

Enable actuator
- Implement database connectivity and memory usage health checks
- Implement successful operations and new users metrics

Implement dev and test environments with different ports and logging levels

+++ src/main/java/com/epam/campus/gymcrm/controllers/impl/TraineeController.java
@@ -1,4 +1,4 @@
-package com.epam.campus.gymcrm.controllers;
+package com.epam.campus.gymcrm.controllers.impl;
 
 import java.util.List;
 import java.util.Map;
@@ -19,6 +19,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.epam.campus.gymcrm.actuator.CustomMetrics;
+import com.epam.campus.gymcrm.controllers.ITraineeController;
 import com.epam.campus.gymcrm.models.dtos.LoginChangeDto;
 import com.epam.campus.gymcrm.models.dtos.LoginDto;
 import com.epam.campus.gymcrm.models.dtos.TraineeCreationDto;
@@ -27,63 +29,51 @@ import com.epam.campus.gymcrm.models.dtos.TraineeUpdateRequestDto;
 import com.epam.campus.gymcrm.models.dtos.TraineeUpdateResponseDto;
 import com.epam.campus.gymcrm.models.dtos.TrainerFromListDto;
 import com.epam.campus.gymcrm.services.impl.TraineeService;
+import com.epam.campus.gymcrm.session.SessionManager;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.persistence.NoResultException;
 import jakarta.persistence.PersistenceException;
 import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/gym-crm/trainees")
-public class TraineeController {
+public class TraineeController implements ITraineeController{
     private static final Logger logger = LoggerFactory.getLogger(TraineeController.class);
     private final TraineeService traineeService;
+    private final CustomMetrics customMetrics;
 
     @Autowired
-    public TraineeController(TraineeService traineeService) {
+    public TraineeController(TraineeService traineeService, CustomMetrics customMetrics) {
         this.traineeService = traineeService;
+        this.customMetrics = customMetrics;
     }
 
     // Login as Trainee
-    @Operation(summary = "Login as Trainee", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Login successful",
-            content = @Content(mediaType = "text/plain")),
-        @ApiResponse(responseCode = "401", description = "Invalid credentials",
-            content = @Content(mediaType = "text/plain"))
-    })
+    @Override
     @GetMapping("/login")
     public ResponseEntity<String> traineeLogin(@Valid @RequestBody LoginDto loginDto) {
         try {
             if (traineeService.traineeLogin(loginDto)) {
+                customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
                 return new ResponseEntity<>("Welcome " + loginDto.getUsername(), HttpStatus.OK);
             } else {
-                return new ResponseEntity<>("Username and password do not match.", HttpStatus.UNAUTHORIZED);
+                return new ResponseEntity<>("Username and password do not match.", HttpStatus.BAD_REQUEST);
             }
 
         } catch (NoSuchElementException e) {
             logger.error(e.getMessage());
-            return new ResponseEntity<>("Username and password do not match. Try again.", HttpStatus.UNAUTHORIZED);
+            return new ResponseEntity<>("Username and password do not match. Try again.", HttpStatus.BAD_REQUEST);
         }
     }
 
     // Create Trainee Profile
-    @Operation(summary = "Create a new Trainee", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Trainee created",
-            content = @Content(schema = @Schema(implementation = LoginDto.class))),
-        @ApiResponse(responseCode = "500", description = "Internal server error",
-            content = @Content)
-    })
+    @Override
     @PostMapping
     public ResponseEntity<LoginDto> createTrainee(@Valid @RequestBody TraineeCreationDto newTraineeDto) {
         try{
             LoginDto loginInfo = traineeService.createTrainee(newTraineeDto);
+            customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
+            customMetrics.addNewUser(); // Increment the new users metric
             return new ResponseEntity<>(loginInfo, HttpStatus.OK);
 
         } catch (PersistenceException e) {
@@ -93,16 +83,16 @@ public class TraineeController {
     }
 
     // Get Trainee profile by username
-    @Operation(summary = "Get trainee profile by username", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Trainee profile found",
-            content = @Content(schema = @Schema(implementation = TraineeResponseDto.class))),
-        @ApiResponse(responseCode = "404", description = "Trainee not found", content = @Content)
-    })
+    @Override
     @GetMapping("/{username}")
     public ResponseEntity<TraineeResponseDto> getTraineeByUsername(@PathVariable String username) {
+        if (!SessionManager.isAuthenticated()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             TraineeResponseDto responseDto = traineeService.getTraineeByUsername(username);
+            customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
             return new ResponseEntity<>(responseDto, HttpStatus.OK);
         } catch (NoSuchElementException e) {
             logger.error("Trainer with that username does not exist");
@@ -111,16 +101,16 @@ public class TraineeController {
     }
 
     // Update Trainee profile
-    @Operation(summary = "Update trainee profile", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Trainee updated",
-            content = @Content(schema = @Schema(implementation = TraineeUpdateResponseDto.class))),
-        @ApiResponse(responseCode = "404", description = "Trainee not found", content = @Content)
-    })
+    @Override
     @PutMapping("/{username}")
     public ResponseEntity<TraineeUpdateResponseDto> updateTrainee(@PathVariable String username, @Valid @RequestBody TraineeUpdateRequestDto traineeUpdateDto) {
+        if (!SessionManager.isAuthenticated()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             TraineeUpdateResponseDto responseDto = traineeService.updateTrainee(username, traineeUpdateDto);
+            customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
             return new ResponseEntity<>(responseDto, HttpStatus.OK);
         } catch (NoSuchElementException e) {
             logger.error("Validation failed and trainee could not be updated: ", e.getMessage());
@@ -129,19 +119,19 @@ public class TraineeController {
     }
 
     // Change Trainee Password
-    @Operation(summary = "Change trainee password", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Password changed", content = @Content(mediaType = "text/plain")),
-        @ApiResponse(responseCode = "401", description = "Unauthorized or invalid password", content = @Content),
-        @ApiResponse(responseCode = "404", description = "Trainee not found", content = @Content)
-    })
+    @Override
     @PutMapping("/{username}/password")
     public ResponseEntity<String> updateTraineePassword(@PathVariable String username, @RequestBody LoginChangeDto loginChangeDto) {
+        if (!SessionManager.isAuthenticated()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             if (traineeService.updateTraineePassword(username, loginChangeDto)) {
+                customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
                 return new ResponseEntity<>("Password changed.", HttpStatus.OK);
             } else {
-                return new ResponseEntity<>("There was an error and password could not be changed.", HttpStatus.UNAUTHORIZED);
+                return new ResponseEntity<>("There was an error and password could not be changed.", HttpStatus.BAD_REQUEST);
             }
 
         } catch (NoSuchElementException | NoResultException e) {
@@ -152,15 +142,16 @@ public class TraineeController {
 
     // Activate/De-activate Trainee
     //! If not idempotent is wanted uncomment this version and comment the following one
-    @Operation(summary = "Switch trainee active status", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Active status changed", content = @Content(mediaType = "text/plain")),
-        @ApiResponse(responseCode = "404", description = "Trainee not found", content = @Content)
-    })
+    @Override
     @PatchMapping("/{username}/active")
     public ResponseEntity<String> switchTraineeActiveStatus(@PathVariable String username) {
+        if (!SessionManager.isAuthenticated()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             boolean setTo = traineeService.switchTraineeActiveStatus(username);
+            customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
             return new ResponseEntity<>("Active status changed to " + setTo, HttpStatus.OK);
         } catch (NoSuchElementException | NoResultException e) {
             logger.error("{} {}", e.getClass(), e.getMessage());
@@ -169,11 +160,17 @@ public class TraineeController {
     }
 
     //! If idempotent is wanted uncomment this version and comment the previous one
+    // @Override
     // @PatchMapping("/{username}/active")
     // public ResponseEntity<String> updateTraineeActiveStatus(@PathVariable String username, @RequestBody Map<String, Object> requestBody) {
+    //     if (!SessionManager.isAuthenticated()) {
+    //         return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+    //     }
+    //
     //     try {
     //         boolean active = Boolean.parseBoolean(requestBody.get("active").toString());
     //         traineeService.updateActiveStatus(username, active);
+    //         customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
     //         return new ResponseEntity<>("Active status updated.", HttpStatus.OK);
     //     } catch (NoSuchElementException | NoResultException e) {
     //         logger.error(e.getMessage());
@@ -182,17 +179,17 @@ public class TraineeController {
     // }
 
     // Update Trainee's trainers list
-    @Operation(summary = "Update trainee's trainers list", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Trainer list updated",
-            content = @Content(schema = @Schema(implementation = TrainerFromListDto.class))),
-        @ApiResponse(responseCode = "404", description = "Trainee not found", content = @Content)
-    })
+    @Override
     @PutMapping("/{username}/trainers")
     public ResponseEntity<List<TrainerFromListDto>> updateTraineeTrainers(@PathVariable String username, @RequestBody Map<String, List<String>> body) {
+        if (!SessionManager.isAuthenticated()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             List<String> newTrainers = body.get("newTrainers");
             List<TrainerFromListDto> trainers = traineeService.updateTraineeTrainersList(username, newTrainers);
+            customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
             return new ResponseEntity<>(trainers, HttpStatus.OK);
         } catch (NoSuchElementException e) {
             logger.error("Validation failed: {}", e.getMessage());
@@ -201,15 +198,16 @@ public class TraineeController {
     }
 
     // Delete Trainee
-    @Operation(summary = "Delete trainee by username", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Trainee deleted", content = @Content(mediaType = "text/plain")),
-        @ApiResponse(responseCode = "404", description = "Trainee not found", content = @Content)
-    })
+    @Override
     @DeleteMapping("/{username}")
     public ResponseEntity<String> deleteTrainee(@PathVariable String username) {
+        if (!SessionManager.isAuthenticated()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         try {
             traineeService.deleteTrainee(username);
+            customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
             return new ResponseEntity<>("Trainee deleted.", HttpStatus.OK);
         } catch (NoSuchElementException | IllegalArgumentException e) {
             logger.error(e.getMessage());
@@ -218,16 +216,16 @@ public class TraineeController {
     }
 
     // Get Trainers list not assigned on trainee, by trainee's username
-    @Operation(summary = "Get list of active trainers not assigned to a trainee", security = @SecurityRequirement(name = "bearerAuth"))
-    @ApiResponses(value = {
-        @ApiResponse(responseCode = "200", description = "Trainers list retrieved",
-            content = @Content(schema = @Schema(implementation = TrainerFromListDto.class))),
-        @ApiResponse(responseCode = "404", description = "Trainee not found", content = @Content)
-    })
+    @Override
     @GetMapping("{traineeUsername}/not-assigned-active-trainers")
     public ResponseEntity<List<TrainerFromListDto>> getTrainersNotAssignedToTrainee(@PathVariable String traineeUsername) {
+        if (!SessionManager.isAuthenticated()) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+        
         try {
             List<TrainerFromListDto> trainers = traineeService.getActiveTrainersNotAssignedToTrainee(traineeUsername);
+            customMetrics.incrementSuccessfulOperation(); // Increment the successful login metric
             return new ResponseEntity<>(trainers, HttpStatus.OK);
 
         } catch (NoSuchElementException e) {
